feat: orchestrator 재고 감소 실패에 대한 보상 트랜잭션 수행

+++ orchestration/src/main/java/com/jylab/dto/response/OrderItem.java
@@ -1,4 +1,4 @@
 package com.jylab.dto.response;
 
-public record OrderItem(Long id, Long productId, Long quantity) {
+public record OrderItem(Long id, Long productId, Integer quantity) {
 }
